🔨 Refactor: nav 스타일 정상화 #66
