+++ src/app/components/index.ts
@@ -1,6 +1,6 @@
 import PcHeader from './header/header_pc'
 import Footer from './footer/footer'
-import Button from './button/button'
+import Button from './button/LinkButton'
 import MobileHeader from './header/header_mobile'
 import TitleForm from './titleForm/titleForm'
 import SectionLayout from './section-layout/SectionLayout'
